fix(condo): DOMA-4109 refetch tickets with comment times
